+++ api/src/main/java/pl/kolendateam/dadcard/race/repository/RegionRepository.java
@@ -5,4 +5,4 @@ import org.springframework.stereotype.Repository;
 import pl.kolendateam.dadcard.race.entity.Region;
 
 @Repository
-public interface RegionRepository extends JpaRepository<Region, Short> {}
+public interface RegionRepository extends JpaRepository<Region, Integer> {}
